// src/components/molecules/Home/EventCard/EventCard.enum.ts
export enum EVENT_CARD_ELEMENTS {
  CARD = 'event-card',
  TITLE = 'event-card-title',
  LOCATION = 'event-card-location',
  DESCRIPTION = 'event-card-description',
  DATE_RANGE = 'event-card-date-range',
  SEVERITY_BADGE = 'event-card-severity-badge',
  TYPE_BADGE = 'event-card-type-badge',
  STATUS_INDICATOR = 'event-card-status-indicator'
}